Backend/requirements.txt
# Crime Tracer Mangalore - Backend Dependencies

# Core Framework
fastapi>=0.100.0
uvicorn>=0.23.0
starlette>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
cryptography>=41.0.0

# Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.24.0
requests>=2.31.0

# AWS S3 (optional)
boto3>=1.28.0

# LLM - Groq (fast inference with Llama models)
groq>=0.5.0
openai>=1.0.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0.0

# Async support
anyio>=3.7.0
greenlet>=2.0.0

# Logging
loguru>=0.7.0

# Rate Limiting
slowapi>=0.1.9
redis>=5.0.0

# File Type Detection
python-magic>=0.4.27

# Monitoring & Metrics
prometheus-client>=0.19.0

# Firebase (for production database)
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0

# Testing (dev dependencies)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Code Quality (dev dependencies)
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# Blockchain (for secure audit trail)
web3>=6.0.0
eth-account>=0.9.0

# ML/AI Models for Forensic Analysis
ultralytics>=8.0.0  # YOLOv8 for object detection
pytesseract>=0.3.10  # Tesseract OCR wrapper
Pillow>=10.0.0  # Image processing for OCR
opencv-python>=4.8.0  # Computer vision support
numpy>=1.24.0  # Numerical operations for image processing
